Text Infrastructure Grows
As the title says, the infrastructure is growing some. Bit tired to fill this out fully, but suffice to say it's just some general progression/bugfixing

+++ GGE2.h
@@ -112,7 +112,7 @@ typedef struct unfinishedObject UnfinObj;
 void defaultMoustClick(GLFWwindow* window, int button, int action, int mods);
 
 //Used for easily creating objects, especially from hardcoded stuff that'd die when touched by realloc
-UnfinObj createUnfinObjFromStatic(float* verts, float* inds, int vLen, int iCount);
+UnfinObj createUnfinObjFromStatic(float* verts, unsigned int* inds, int vLen, int iCount);
 
 //Appends objTwo into objOne. Needs both to have their arrays dynamically allocated
 void appendUnfinisheds(UnfinObj* objOne, UnfinObj* objTwo);
